test: . move tests of field rendering
